cleanup #5

+++ VulkanEngine/Level.cpp
@@ -21,7 +21,7 @@ void Level::InitializeLevel(const VkCommandPool& commandPool, const glm::mat4& p
     m_3DUBOMatrixes.proj = projMatrix;
 
     m_2DUBOMatrixes.model = glm::mat4(1.f);
-    float aspectRatio = static_cast<float>(VulkanBase::swapChainExtent.width) / VulkanBase::swapChainExtent.height;
+    float aspectRatio = static_cast<float>(VulkanBase::swapChainExtent.width) / static_cast<float>(VulkanBase::swapChainExtent.height);
     glm::vec3 scaleFactors(1 / aspectRatio, 1, 1.0f);
     m_2DUBOMatrixes.view = glm::scale(glm::mat4(1.0f), scaleFactors);
 
@@ -67,7 +67,7 @@ void Level::Draw2DMeshes(const VkCommandBuffer& commandBuffer, const uint32_t cu
 void Level::WindowHasBeenResized(const glm::mat4& projMatrix)
 {
     m_3DUBOMatrixes.proj = projMatrix;
-    float aspectRatio = static_cast<float>(VulkanBase::swapChainExtent.width) / VulkanBase::swapChainExtent.height;
+    float aspectRatio = static_cast<float>(VulkanBase::swapChainExtent.width) / static_cast<float>(VulkanBase::swapChainExtent.height);
     glm::vec3 scaleFactors(1 / aspectRatio, 1, 1.0f);
     m_2DUBOMatrixes.view = glm::scale(glm::mat4(1.0f), scaleFactors);
 }
